Prevent default submission in contact page button
As of right now, the form doesn't do anything with the information. We
could leave it, but upon press, it causing the page to reload, and as of
right now, I'd like to not have that happen until I can come up with a
fun way to use the form info (maybe figured out in a later lessons). For
authenticating user input and verifying submission details, I should be
sure to reinstate the submission behavior.