Remove dist from gitignore, create new page
I had some trouble working around untracking dist and then retracking it
when I want my pdeployed page to reflect the updated repository changes.
I deleted the ghpages branch and started fresh with a new deployment. I
was running into issues with certain pull and push requests not working
with my gh-page branch, so I decided to clear it and look deeper into
this question when I come back to it.